requirements.txt
streamlit>=1.30.0
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
